+++ CHANGELOG.rst
@@ -1,3 +1,11 @@
+UNRELEASED
+----------
+
+* Fix type annotations:
+  * ``oop_ext.foundation.weak_ref.IsWeakProxy``.
+  * ``oop_ext.foundation.weak_ref.IsWeakRef``.
+  * ``oop_ext.foundation.weak_ref.GetWeakProxy``.
+
 2.2.0
 -----